Added project_root to the credentials
Now you can specify the root in the credentials, or you
can specify an environment variable you'd like to use.
Later shopifydev will just predict what to use, but for
now since I don't know the right ENV variable to use
in VIM... we'll do it this way!

Also, documented that.